+++ tests/Server/ScriptTest.php
@@ -9,11 +9,11 @@ use Innmind\Server\Control\{
     Server\Command,
     Exception\ProcessFailed,
 };
-use Innmind\TimeContinuum\{
+use Innmind\Time\{
     Clock,
     Period,
+    Halt,
 };
-use Innmind\TimeWarp\Halt;
 use Innmind\IO\IO;
 use Innmind\Immutable\SideEffect;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;