+++ tests/Server/VolumesTest.php
@@ -11,8 +11,10 @@ use Innmind\Server\Control\{
     Server\Command,
     Exception\ProcessFailed,
 };
-use Innmind\TimeContinuum\Clock;
-use Innmind\TimeWarp\Halt;
+use Innmind\Time\{
+    Clock,
+    Halt,
+};
 use Innmind\IO\IO;
 use Innmind\Url\Path;
 use Innmind\Immutable\SideEffect;