+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `Innmind\Server\Control\Server` is now a final class, all previous implementations are now flagged as internal
 - `Innmind\Server\Control\Server\Volumes` is now a final class
 - `Innmind\Server\Control\Server\Processes` is now a final class
+- Requires `innmind/time:~1.0`
 
 ## 6.1.0 - 2025-08-06